requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=1.5.0
